+++ libp2p/network/url_conversion_test.go
@@ -36,7 +36,7 @@ func TestPeerMultiAddrFromServiceURL_Success(t *testing.T) {
 		},
 		{
 			"IPv6",
-			"2a00:1450:4005:802::2004:8080",
+			"[2a00:1450:4005:802::2004]:8080",
 			"/ip6/2a00:1450:4005:802::2004/tcp/8080",
 		},
 	}
@@ -63,40 +63,37 @@ func TestPeerMultiAddrFromServiceURL_Error(t *testing.T) {
 	hostnames := map[string]string{
 		fqdn: "www.google.com",
 		ip4:  "142.250.181.196",
-		ip6:  "2a00:1450:4005:802::2004",
+		ip6:  "[2a00:1450:4005:802::2004]",
 	}
 
 	testCases := []struct {
 		name             string
 		serviceURLFormat string
-		// TECHDEBT: assert specific errors.
-		expectedErrContains string
+		expectedErr      string
 	}{
 		// Usage of scheme is invalid.
 		{
 			"fully qualified domain name with scheme",
 			"tcp://%s:8080",
-			"resolving peer IP for hostname",
+			errInvalidSchemeUsageMsg,
 		},
 
 		// Port **number** is required
 		{
 			"invalid port number",
 			"%s:abc",
-			"invalid port",
+			errParsingServiceURLMsg,
 		},
 		{
 			"missing port number",
 			"%s:",
-			"unexpected end of multiaddr",
+			errMissingPortAndDelimiterMsg,
 		},
-		// TODO: this case is tricky to detect as IPv6 addresses
-		// can omit multiple "hextet" delimiters and still be valid.
-		// (see: https://en.wikipedia.org/wiki/IPv6#Address_representation)
-		// {
-		// 	"missing port number and delimiter",
-		// 	"%s",
-		// },
+		{
+			"missing port number and delimiter",
+			"%s",
+			errMissingPortAndDelimiterMsg,
+		},
 	}
 
 	for _, testCase := range testCases {
@@ -104,13 +101,33 @@ func TestPeerMultiAddrFromServiceURL_Error(t *testing.T) {
 			testName := fmt.Sprintf("%s/%s", testCase.name, hostType)
 			t.Run(testName, func(t *testing.T) {
 				serviceURL := fmt.Sprintf(testCase.serviceURLFormat, hostname)
-				actualMultiaddr, err := Libp2pMultiaddrFromServiceURL(serviceURL)
-				// TECHDEBT: assert specific errors
-				// Print resulting multiaddr to understand why no error.
-				require.ErrorContainsf(t, err, testCase.expectedErrContains, fmt.Sprintf("actualMultiaddr: %s", actualMultiaddr))
+				_, err := Libp2pMultiaddrFromServiceURL(serviceURL)
+				require.ErrorContainsf(t, err, testCase.expectedErr, "unexpected error")
 			})
 		}
 	}
+}
+
+func ExampleLibp2pMultiaddrFromServiceURL() {
+	// Example: IPv4
+	peerMultiAddr, err := Libp2pMultiaddrFromServiceURL("142.251.40.164:8080")
+	if err != nil {
+		panic(err)
+	}
+	fmt.Println(peerMultiAddr)
+
+	// Output: /ip4/142.251.40.164/tcp/8080
+}
+
+func ExampleLibp2pMultiaddrFromServiceURL_ipv6() {
+	// Example: IPv6
+	peerMultiAddr, err := Libp2pMultiaddrFromServiceURL("[2a00:1450:4005:802::2004]:8080")
+	if err != nil {
+		panic(err)
+	}
+	fmt.Println(peerMultiAddr)
+
+	// Output: /ip6/2a00:1450:4005:802::2004/tcp/8080
 }
 
 func TestServiceURLFromLibp2pMultiaddr_Success(t *testing.T) {
@@ -127,18 +144,18 @@ func TestServiceURLFromLibp2pMultiaddr_Success(t *testing.T) {
 		{
 			"IPv6 full",
 			"/ip6/2a00:1450:4005:0802:0000:0000:0000:2004/tcp/8080",
-			"2a00:1450:4005:802::2004:8080",
+			"[2a00:1450:4005:802::2004]:8080",
 		},
 		{
 			"IPv6 short",
 			"/ip6/2a00:1450:4005:802::2004/tcp/8080",
-			"2a00:1450:4005:802::2004:8080",
+			"[2a00:1450:4005:802::2004]:8080",
 		},
 		{
 			"IPv6 shorter",
 			// NB: this address is not equivalent to those above.
 			"/ip6/2a00::2004/tcp/8080",
-			"2a00::2004:8080",
+			"[2a00::2004]:8080",
 		},
 	}