+++ app/client/doc/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.23] - 2023-03-21
+
+- Refactor debug CLI to use new P2P interfaces
+
+## [0.0.0.22] - 2023-03-17
+
+- Added a limit on concurrent key imports for debug client to avoid OOM.
+
 ## [0.0.0.21] - 2023-03-14
 
 - Simplifies the debug CLI tooling by embedding private-keys.yaml manifest